Object_Face_Detection/requirements.txt
flask
werkzeug
opencv-python
numpy
Pillow
ultralytics
supervision
requests
streamlit
face-recognition
qrcode
python-dotenv 